#162 merge day_by_day transactions
