--- compute_transactions/function.json
@@ -1,10 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "name": "payload",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ]
-}
